refactor(blorktools): modularizing
